Added code samples, integration tests to support 2.9
Added 3 code samples
1. basic
2. custom alias file
3. custom alias file resolver
Changed the way integration tests work to support gradle 2.9
to test a settings plugin like this one we needed to use
gradle testkit which is still in incubation meaning it has not
backwards compatability. the test as they were support 2.6-2.8.
with 2.9 out we needed to support both version by inspecting the
metaclass.